requirements.txt
certifi==2022.5.18.1
chardet==4.0.0
click==8.1.3
dominate==2.6.0
Flask==2.1.2
Flask-Bootstrap4==4.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
requests==2.27.1
SQLAlchemy==1.4.37
urllib3==1.26.9
visitor==0.1.3
Werkzeug==2.1.2
WTForms==3.0.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
Flask-Gravatar==0.5.0
Flask-Login==0.6.1
numpy==1.22.4
stripe==2.27.1